+++ src/Routes/api.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 use Werify\Account\Laravel\Http\Controllers\V1\Auth\ClassicController;
+use Werify\Account\Laravel\Http\Controllers\V1\Authorize\ClassicController as ClassicControllerAlias;
 
 Route::prefix(config('waccount.routes.prefix'))->group(function () {
 
@@ -12,6 +13,14 @@ Route::prefix(config('waccount.routes.prefix'))->group(function () {
             Route::post('register', 'register')->name('register');
             Route::post('me', 'me')->name('me');
             Route::post('logout', 'logout')->name('logout');
+        });
+    });
+
+    // Authorize
+    Route::prefix('authorize')->name('authorize.')->group(function(){
+        Route::prefix('classic')->name('classic.')->controller(ClassicControllerAlias::class)->group(function(){
+            Route::post('start', 'start')->name('start');
+            Route::post('check', 'check')->name('check');
         });
     });